requirements.txt
numpy==1.26.4
openai==1.30.1
python-dotenv==1.0.1
tiktoken==0.7.0
flask==3.0.3
opencv-python==4.9.0.80
uuid==1.30